LayoutTests/fast/speechsynthesis/speech-synthesis-crash-on-bad-utterance.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="console"></div>

<script>

    description("This tests that passing in the wrong type of data won't crash speech synthesis code");

     // Don't crash. Speak is supposed to take an utterance, not a string.
     shouldThrow("speechSynthesis.speak('Hello World')");

     // Don't crash. An utterance voice is supposed to take a voice object, not a string.
     var x = new SpeechSynthesisUtterance('Hello World');
     shouldThrow('x.voice = "asdf"');

     // Don't crash. An utterance is supposed to take a string, not a number.
     x = new SpeechSynthesisUtterance(223);

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
